feat(absorbedBugsPanel): organizes AbsorbedBugItem component, integrates absorbedBugs arrays from server and update styles
