--- client/src/components/AbsorbedBugItem/AbsorbedBugItem.module.css
@@ -1,15 +1,0 @@
-.AbsorbedBugItemWrapper {
-  width: 95%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 0.6rem;
-  margin: 0.1rem 0;
-}
-.AbsorbedBugItemLabel {
-  color: #ffffff;
-  font-size: var(--font-md);
-  font-family: var(--font-family);
-  text-wrap: nowrap;
-}
