+++ client/src/components/absorbedBugsPanel/AbsorbedBugItem.jsx
@@ -1,9 +1,9 @@
 // eslint-disable-next-line no-unused-vars
-import styles from './AbsorbedBugItem.module.css';
+import styles from './AbsorbedBugsPanel.module.css';
 
 function AbsorbedBugItem({ component }) {
   return (
-    <div className={styles.AbsorbedBugItemWrapper}>
+    <div className={`${styles.AbsorbedBugItemWrapper} ${styles.updated}`}>
       <span className={styles.AbsorbedBugItemLabel}>{component}</span>
     </div>
   );
